refactor(decompose): split webapp/e2e-tests/electron/for_feature_development_not_LT_verification/editors/electron-markdown-editor-yaml-integrity.spec.ts into folder
