+++ webapp/e2e-tests/electron/for_feature_development_not_LT_verification/editors/electron-markdown-editor-yaml-integrity.spec.ts
@@ -15,189 +15,10 @@
  * IMPORTANT: THESE SPEC COMMENTS MUST BE KEPT UP TO DATE
  */
 
-import { test as base, expect, _electron as electron } from '@playwright/test';
-import type { ElectronApplication, Page } from '@playwright/test';
 import * as path from 'path';
 import * as fs from 'fs/promises';
-import * as os from 'os';
-import type { Core as CytoscapeCore } from 'cytoscape';
-import type { EditorView } from '@codemirror/view';
-
-// Use temp directory for this test
-const PROJECT_ROOT = path.resolve(process.cwd());
-
-// Type definitions
-interface ExtendedWindow {
-  cytoscapeInstance?: CytoscapeCore;
-  electronAPI?: {
-    main: {
-      startFileWatching: (dir: string) => Promise<{ success: boolean; directory?: string; error?: string }>;
-      stopFileWatching: () => Promise<{ success: boolean; error?: string }>;
-      getWatchStatus: () => Promise<{ isWatching: boolean; directory?: string }>;
-      getGraph: () => Promise<{ nodes: Record<string, unknown> } | undefined>;
-    };
-  };
-}
-
-// Helper type for CodeMirror access
-interface CodeMirrorElement extends HTMLElement {
-  cmView?: { view: EditorView };
-}
-
-// Extend test with Electron app
-const test = base.extend<{
-  electronApp: ElectronApplication;
-  appWindow: Page;
-  testVaultPath: string;
-}>({
-  // Create temp userData directory with embedded vault + config
-  // The config auto-loads the vault during app initialization
-  // IMPORTANT: Files must be in {watchedFolder}/voicetree/ due to default vaultSuffix
-  electronApp: [async ({}, use, testInfo) => {
-    // Create temp userData directory
-    const tempUserDataPath = await fs.mkdtemp(path.join(os.tmpdir(), 'voicetree-yaml-integrity-test-'));
-
-    // Create the watched folder (what config points to)
-    const watchedFolder = path.join(tempUserDataPath, 'test-vault');
-    await fs.mkdir(watchedFolder, { recursive: true });
-
-    // Create the actual vault path with default suffix 'voicetree'
-    // The app looks for .md files in {watchedFolder}/voicetree/
-    const projectRoot = path.join(watchedFolder, 'voicetree');
-    await fs.mkdir(projectRoot, { recursive: true });
-
-    // Create test files for both tests
-    const testNodeId = 'test-yaml-node.md';
-    const testNodeId2 = 'test-yaml-preservation.md';
-
-    const initialContent = `---
-node_id: 42
-color: "#FF5733"
-tags:
-  - important
-  - test
-customField: some value
----
-# My Important Node Title
-
-This is the content of my node.
-
-Some additional text here.`;
-
-    const initialContent2 = `---
-node_id: 123
-color: "#00FF00"
-position:
-  x: 100
-  y: 200
-isContextNode: true
-customArray:
-  - item1
-  - item2
-customObject:
-  key1: value1
-  key2: value2
----
-# Preserved Title
-
-Original content here.`;
-
-    await fs.writeFile(path.join(projectRoot, testNodeId), initialContent, 'utf-8');
-    await fs.writeFile(path.join(projectRoot, testNodeId2), initialContent2, 'utf-8');
-
-    // Write config to auto-load the watched folder (vault = watchedFolder + 'voicetree')
-    const configPath = path.join(tempUserDataPath, 'voicetree-config.json');
-    await fs.writeFile(configPath, JSON.stringify({ lastDirectory: watchedFolder }, null, 2), 'utf8');
-    console.log('[Test] Watched folder:', watchedFolder);
-    console.log('[Test] Vault path (with suffix):', projectRoot);
-
-    // Store projectRoot for test access via testInfo (the actual path where .md files live)
-    (testInfo as unknown as { projectRoot: string }).projectRoot = projectRoot;
-
-    const electronApp = await electron.launch({
-      args: [
-        path.join(PROJECT_ROOT, 'dist-electron/main/index.js'),
-        `--user-data-dir=${tempUserDataPath}` // Isolate test userData
-      ],
-      env: {
-        ...process.env,
-        NODE_ENV: 'test',
-        HEADLESS_TEST: '1',
-        MINIMIZE_TEST: '1'
-      },
-      timeout: 8000
-    });
-
-    await use(electronApp);
-
-    // Graceful shutdown
-    try {
-      const page = await electronApp.firstWindow();
-      await page.evaluate(async () => {
-        const api = (window as ExtendedWindow).electronAPI;
-        if (api) {
-          await api.main.stopFileWatching();
-        }
-      });
-      await page.waitForTimeout(300);
-    } catch {
-      console.log('Note: Could not stop file watching during cleanup');
-    }
-
-    await electronApp.close();
-
-    // Cleanup entire temp directory (includes vault)
-    await fs.rm(tempUserDataPath, { recursive: true, force: true });
-    console.log('[Test] Cleaned up temp directory');
-  }, { timeout: 45000 }],
-
-  // Get vault path from testInfo (set by electronApp fixture)
-  testVaultPath: async ({}, use, testInfo) => {
-    // Wait for electronApp fixture to set projectRoot
-    await use((testInfo as unknown as { projectRoot: string }).projectRoot);
-  },
-
-  appWindow: [async ({ electronApp, testVaultPath: _testVaultPath }, use) => {
-    const page = await electronApp.firstWindow();
-
-    // Log console messages
-    page.on('console', msg => {
-      console.log(`BROWSER [${msg.type()}]:`, msg.text());
-    });
-
-    page.on('pageerror', error => {
-      console.error('PAGE ERROR:', error.message);
-      console.error('Stack:', error.stack);
-    });
-
-    await page.waitForLoadState('domcontentloaded', { timeout: 30000 });
-
-    // Check for errors before waiting for cytoscapeInstance
-    const hasErrors = await page.evaluate(() => {
-      const errors: string[] = [];
-      // Check if React rendered
-      if (!document.querySelector('#root')) errors.push('No #root element');
-      // Check if any error boundaries triggered
-      const errorText = document.body.textContent;
-      if (errorText?.includes('Error') || errorText?.includes('error')) {
-        errors.push(`Page contains error text: ${errorText.substring(0, 200)}`);
-      }
-      return errors;
-    });
-
-    if (hasErrors.length > 0) {
-      console.error('Pre-initialization errors:', hasErrors);
-    }
-
-    await page.waitForFunction(() => (window as ExtendedWindow).cytoscapeInstance, { timeout: 30000 });
-
-    // Wait for electronAPI to be available
-    await page.waitForFunction(() => (window as ExtendedWindow).electronAPI?.main, { timeout: 30000 });
-    await page.waitForTimeout(500); // Give extra time for auto-load to complete
-
-    await use(page);
-  }, { timeout: 30000 }]
-});
+import { expect, test } from './electron-markdown-editor-yaml-integrity/test-fixtures';
+import type { CodeMirrorElement, ExtendedWindow } from './electron-markdown-editor-yaml-integrity/test-fixtures';
 
 test.describe('Markdown Editor YAML Integrity', () => {
   test('should NOT corrupt YAML frontmatter when editing markdown in floating editor', async ({ appWindow, testVaultPath }) => {
